Code cleanup.

+++ src/qw/cl_cam.c
@@ -81,12 +81,12 @@ vectoangles (vec3_t vec, vec3_t ang)
 		else
 			pitch = 270;
 	} else {
-		yaw = (int) (Q_atan2 (vec[1], vec[0]) * 180 / M_PI);
+		yaw = Q_atan2 (vec[1], vec[0]) * 180 / M_PI;
 		if (yaw < 0)
 			yaw += 360;
 
 		forward = Q_sqrt (vec[0] * vec[0] + vec[1] * vec[1]);
-		pitch = (int) (Q_atan2 (vec[2], forward) * 180 / M_PI);
+		pitch = Q_atan2 (vec[2], forward) * 180 / M_PI;
 		if (pitch < 0)
 			pitch += 360;
 	}
